Add files via upload
# BYJU's Login Page Clone

Welcome to the BYJU's Login Page Clone repository! 🚀

## Overview

This repository contains the HTML and CSS code for a faithful reproduction of BYJU's login page. Whether you're a developer looking to contribute, a learner eager to understand web development, or someone interested in exploring BYJU's design, this project is a great place to start.

## Features

- **Responsive Design:** Ensures a seamless user experience across various devices and screen sizes.
- **CSS Styling:** Leverages modern CSS techniques for a polished and visually appealing interface.
- **Easy to Understand:** Codebase is well-commented and structured, making it beginner-friendly and educational.

## Getting Started

1. Clone the repository: `https://github.com/ut-stack/Byju-s-Login-Clone.git`
2. Open the `index.html` file in your preferred web browser.

Feel free to explore, modify, and enhance the code to suit your learning or development goals. Contributions are always welcome!

## Contributing

If you'd like to contribute, please fork the repository and create a pull request. Issues and feature requests can be submitted in the  section https://github.com/ut-stack/Byju-s-Login-Clone.git .

## Credits

This project is maintained by Utkarsh Tripathi. Special thanks to BYJU's for the inspiration.

Happy coding! 🌐✨

---